requirements.txt
discord.py
git+https://github.com/Gorialis/jishaku
asyncpg
python-dotenv
better_exceptions
aiohttp
tabulate